Refactor publisher into its own module
